change int n ---> long n, for long inputs

+++ cpp/fel/HW02/b3b36prg-hw02/main.c
@@ -7,9 +7,9 @@
 #define primes_in_mil 78498 // count of prime numbers in milion
 
 int read_number(void);
-int decompose(int n, int primes[]);
+int decompose(long n, int primes[]);
 void prime_numbers(int primes[]);
-int print_output(int counter, int i, int actual_num, int n, int primes[]);
+int print_output(int counter, int i, int actual_num, long n, int primes[]);
 
 // solution 1 // mandatory
 // deviding input by smaller number than input number
@@ -25,7 +25,7 @@ int main(int argc, char *argv[])
 {
     int primes[primes_in_mil];
     int ret = EXIT_SUCCESS;
-    int n;
+    long n;
     while ((n = read_number()) > 0) {
         printf("Prvociselny rozklad cisla %d je:\n", n);
         prime_numbers(primes);
@@ -40,7 +40,7 @@ int main(int argc, char *argv[])
 
 int read_number(void)
 {
-    int n = -1;
+    long n = -1;
     if (scanf("%d", &n) != 1) {
         n = -1;
     }
@@ -62,7 +62,7 @@ void prime_numbers(int primes[])
     }
 }
 
-int decompose(int n, int primes[])
+int decompose(long n, int primes[])
 {
     int counter = 0; // counting exponent
     int actual_num = n;
@@ -86,7 +86,7 @@ int decompose(int n, int primes[])
     return EXIT_SUCCESS;
 }
 
-int print_output(int counter, int i, int actual_num, int n, int primes[])
+int print_output(int counter, int i, int actual_num, long n, int primes[])
 {
     // condition true only if input number is 1
     if (n == 1) {
